requirements.txt
psycopg2-binary==2.9.9
python-dotenv==0.19.1
Flask==3.0.3
gunicorn==21.2.0
Werkzeug==3.0.2

